refactor(epub-preview): extract render cache and asset modules
